wms-dataflow/app/requirements.txt
mysql-connector-python==8.0.24
OdooRPC==0.7.0
